requirements.txt
numpy
gymnasium
gymnasium-robotics
gymnasium[other]
stable-baselines3[extra]
huggingface_hub
openai
python-dotenv

# for running training scripts of model
# torch
# wandb

